chore(ComponentExample): support React live demo
This change adds React live demo and code snippet feature to
`<ComponentExample>`.

* The React live demo feature is implemented as
  `<ComponentExampleLiveReact>`, which has the same API as its vanilla
  conterpart (`<ComponentExampleLive>`) and thus can be swapped back
  and forth by user's selection of JavaScript framework variants for
  demos
* The React code snippet feature is implemented as
  `<CodeExampleReact>`, which has the same API as its plain HTML
  counterpart (`<CodeExampleHTML>`) and thus can be swapped back and
  forth by user's selection of JavaScript framework variants for demos

The React live demo feature is enabled with `hasReactVersion="live"`
prop of `<ComponentCode>`. It can be also set as the default by
`defaultFrameworkVariant="react"` prop of `<ComponentCode>`. If
`defaultFrameworkVariant="react"` is set, vanilla live demo won't be
enabled unless `hasVanillaVersion={true}` prop of `<ComponentCode>` is
set.

The source of the live demo as well as of the code snippet comes from a
map of React components, keyed by the component name and the variant
name specified for `<ComponentCode>`, which is located at
`src/react-demo/index.js`.

The code snippet for the React version is generated as a conversion
from a shallow React component tree to a JSX string. Due to its
"shallow" nature, using `<DemoComponent foo={fooValue}>` for the
React-component-tree-to-JSX converter just emits
`<DemoComponent foo={fooValue}>` (as the code snippet), and thus we run
`DemoComponent({ foo: fooValue })` instead to pass along to the
converter to get the JSX of the React component tree of what
`DemoComponent` generates. Those demo components has to be implemented
as pure functions therefore.

This change also removes drop-up demo for now due to the limited
availablity of the demo.

Refs carbon-design-system/carbon#1936.