chore(ComponentReact): use it for entire dropdown page
This change is for using React live demo feature for the entire
dropdown page, to demonstrate the correct selection feature in
dropdown. The React live demo feature is enhanced to support the
dropdown component, as well as to support vanilla/Vue links.

This is a stop-gap change until the team refactors the live demo
feature in the website as a whole.

Refs carbon-design-system/carbon#1936.