Clarify Use Of React Web Hooks Definition
The instructions, as originally written, are  ambiguous because there are multiple `return()` statements in the `RepoPage` component. My change clarifies this by specifying that the React hook configuration needs to be done before the outer or first `return`.  Additionally, I show that they need to be declared at the top of the function.

If the React hooks are placed before the inner `return()` statement, a somewhat opaque error message is given: "Rendered more hooks than during the previous render." Understanding this requires the reader to dig into the React hooks documentation.