fix(dropdown): fix example Dropdown data object (#1265)
When using 'text', no text shows up for the options listed. Changing each one to 'label' makes them show up in the dropdown. I'm not sure why, or where to even look for further documentation on this, and changing it to 'label' was just a wild guess that ended up working.

Co-authored-by: TJ Egan <tw15egan@gmail.com>
Co-authored-by: Josh Black <josh@josh.black>