feat: add on demand and scheduled image processing
This updates the image processing to allow for running on-demand and scheduled (once a week on Sunday). This prevents the bot from cluttering (sometimes unrelated) pull requests, which has been a source of confusion during code reviews.

After manually triggering the action in a fork, the bot successfully opened a PR shaving off 10mb from the images.