Replace deprecated layout token with spacing token
One of the code examples uses the scss variables `$layout-05` for vertical padding/spacing. 
No mention of the layout token system is made in the tutorial (only the spacing tokens are referenced). 
It seems that the layout tokens are deprecated, and it's a bit confusing for tutorial users to see a rogue example of the layout token with no context or explanation.
Suggest replacing it with $spacing-10 (equivalent rem).