Update usage.mdx
Removed screen reader info; usual editorial tweaks with Keyboard. Made "group label" the consistent name by removing "heading" (as discussed).
Most important change to scrutinize: the repositioning of the 2nd bullet of the label to the group label area. Took a guess at what was meant; may not be correct.