docs(modal): Update usage.mdx (#2868)
* Update usage.mdx

Some suggested changes, to be discussed at the Guild call.

* Update usage.mdx

matched existing language

* Update usage.mdx

* Update usage.mdx

Altered "clear" to be "delete". Delete is the better option, and I only chose clear because of an implementation example. But since there is also an example using 'delete', we'll stick with the usual wording.

* Update usage.mdx

As discussed with Lauren Rice, updating the Anatomy diagram and description to add the Delete button

* Update usage.mdx

added formatting

* docs(file-uploader): run prettier

* Update file-uploader-usage-1.png

Replacing image supplied by @laurenmrice  to match text update

* Update style.mdx

Altered label to heading to match usage update, as per @aagonzales request

* fix(website): run format

Co-authored-by: Josh Black <josh@josh.black>
Co-authored-by: Abbey Hart <abbeyhrt@gmail.com>