File uploader accessibility (#2886)
* Update to accessibility tab

Added images and rewrote content

* Update usage.mdx

Added link to accessibility tab from Usage

* Update accessibility.mdx

Added not entirely resolved issues for Carbon. I'd like these to be specific -- and for the page to capture how Carbon accomplishes. But until the bugs are addressed, this identifies the consideration.

* Update accessibility.mdx

fixed typos

* Update accessibility.mdx

trying to put in a line break to pass the checker...

* updating images and ALT text

* Update accessibility.mdx

changes list style in Considerations

* you're to you are