Update step-1.mdx (#3071)
* Update step-1.mdx

Hi,
it took me quiet some time to run the step-1 part of the tutorial, the section about icons namely.
After digging into documentation of `carbon-component-angular` module I was able to make the icons working properly.
Therefore you would like to propose couple changes in this part of the step-1 documentation.

* Update step-1.mdx

Forgot to mention the registration of the icon/s so adding that part as well.

* Update src/pages/developing/angular-tutorial/step-1.mdx

ok, thx

Co-authored-by: Akshat Patel <38994122+Akshat55@users.noreply.github.com>

* Update src/pages/developing/angular-tutorial/step-1.mdx

ok thx

Co-authored-by: Akshat Patel <38994122+Akshat55@users.noreply.github.com>

* Update src/pages/developing/angular-tutorial/step-1.mdx

ok thx

Co-authored-by: Akshat Patel <38994122+Akshat55@users.noreply.github.com>

Co-authored-by: Akshat Patel <38994122+Akshat55@users.noreply.github.com>
Co-authored-by: D.A. Kahn <40970507+dakahn@users.noreply.github.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>