Corrections from manual check
A number of pages had the A11yStatus formed with indented code, so my GSR failed to locate.
As well, a manual check of each page (which I did by navigating to the Accessibility tab and clicking the Development considerations on-page link), showed that a few of the on-page links were malformed.