Update accessibility.mdx (#4184)
* Update accessibility.mdx

Tried moving the accessibility status to the bottom of the page, as per #3960 

Since the H3 is generated in code, I'm just leaving it as is. Expectation is that this will be a subsection of Development considerations, which works fine.
Not showing up in preview.

* Move accessibility testing to bottom of tabbed content

Changed 28 components that were previously revised to the standard accessibility tab layout

* Relocate testing on non-standard pages

Move the accessibility testing to the bottom of pages that have not been updated to the current accessibility tab template

* Corrections from manual check

A number of pages had the A11yStatus formed with indented code, so my GSR failed to locate.
As well, a manual check of each page (which I did by navigating to the Accessibility tab and clicking the Development considerations on-page link), showed that a few of the on-page links were malformed.

* Update accessibility.mdx

Missed popover testing section

* Removed old testing information

These five components had not had their Accessibility tab info updated due to various reasons. We elected to strip out the old testing results.
I left the H2 for Testing in place.

* Update accessibility.mdx

Adding top description to list component.