Edits on usage tab - content section - bullet on on current page
Edits on Usage tab

Content section:

For this bullet in this section: “By default, the current page is not listed in the breadcrumb trail.”:

Since all of the images on our website are now showing the current page in the trail, this bullet can be expanded on to explain this more. Can you make this bullet more descriptive to cover more of what we say in the Modifiers section.