Usage tab - modifiers
Changes to text to read "By default, Carbon breadcrumb trails should include the current page. If a page doesn’t have a title or the current page is not clear, it can be included in the breadcrumb trail. If the current page is included in a breadcrumb trail, it is always the last text listed and is not an interactive link."