Update to style image - spec only 16px and 18px removing rem
From comment:

After seeing this Size image live in the website preview, you can just annotate 16px and 18px in magenta and remove the rem part in the image. For most of our Size section images on the website, we don't include the rem here, so just following this for consistency.