Docs: Breadcrumb [small breadcrumb] website updates (#4514)
* Breadcrumb doc updates

Changes:
-Sizing image
ect..

* Structure update

* Updates

* Updates to usage tab

* Updates style tab

* Updates Medium breadcrumb

* Updates to style tab

* Updates Accessibility

* Captions (periods or no periods)

* Edits on usage tab - content section - bullet on on current page

Edits on Usage tab

Content section:

For this bullet in this section: “By default, the current page is not listed in the breadcrumb trail.”:

Since all of the images on our website are now showing the current page in the trail, this bullet can be expanded on to explain this more. Can you make this bullet more descriptive to cover more of what we say in the Modifiers section.

* Usage tab - modifiers

Changes to text to read "By default, Carbon breadcrumb trails should include the current page. If a page doesn’t have a title or the current page is not clear, it can be included in the breadcrumb trail. If the current page is included in a breadcrumb trail, it is always the last text listed and is not an interactive link."

* Updates - Style tab - Color section - simplified to one image with overflow

Updates - Style tab
Color section:

After seeing this live in the website preview, I think instead of having two image in this section, you can just keep the breadcrumb image that also have the overflow to keep things more concise. The two images look a little redundant.

* Update to image - truncation

Re-uploaded image to avoid pixilation and the boarder around caret

* Update src/pages/components/breadcrumb/style.mdx

Co-authored-by: Lauren Rice <43969356+laurenmrice@users.noreply.github.com>

* Updates

A group of breadcrumbs with overflow content uses the overflow menu component to disclose additional breadcrumbs in a menu view. A tooltip component appears in the hover, focus, and active states. Use the [menu](/components/menu/style) and [tooltip](/components/tooltip/style) colors for these instances.

* Updates $icon-primary

* Update to structure - with correct code for images

<div className="image--fixed">

![Structure and spacing measurements for a checkbox group](images/checkbox-style-3.png)

</div>

<Caption>
  Structure and spacing measurements for a checkbox group | px / rem
</Caption>

* Update to style image - spec only 16px and 18px removing rem

From comment:

After seeing this Size image live in the website preview, you can just annotate 16px and 18px in magenta and remove the rem part in the image. For most of our Size section images on the website, we don't include the rem here, so just following this for consistency.

* Updates - Usage tab - Content section

- Bullet content change " By default, the current page is not listed in the breadcrumb trail. However, if a page doesn’t have a title or the current page is not clear, the current page can be included in the breadcrumb trail if it is more suitable for your products use case. If the current page is included in a breadcrumb trail, it is always the last text listed and is not an interactive link."

- UI Images re-uploaded removing current breadcrumb "UI images on the Usage tab that have page titles, you can remove the current page breadcrumb, but keep the slash after the last blue breadcrumb and re-upload those images so it follows our guidance. (I already updated those images for you in your file)"

- Modifiers section content update " By default, Carbon breadcrumb trails should not include the current page. If a page doesn’t have a title or the current page is not clear, it can be included in the breadcrumb trail. If the current page is included in a breadcrumb trail, it is always the last text listed and is not an interactive link."

* Re-upload image at 4X

* chore: yarn format

---------

Co-authored-by: Lauren Rice <43969356+laurenmrice@users.noreply.github.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>
Co-authored-by: Taylor Jones <taylor.jones826@gmail.com>